Extract store logic outside of index.js, so that it can be used in other parts of the app (you cannot set index.js (entry point to webpack) as a dependency in any other module)
