+++ client/reducers/index.js
@@ -8,6 +8,18 @@ import view from './view'
 import user from './user'
 import { routerReducer } from 'react-router-redux';
 
+
+
+const routeFix = () => {
+  // rehydrate state here
+
+  // return routerReducer here passing
+
+
+}
+
+
+
 const reducers = combineReducers({
   auth,
   // grid,
